requirements.txt
# Minimal runtime deps (expand later)
opencv-python
numpy
pydantic
pyyaml
prometheus-client
